requirements.txt
Django==4.2.7
djangorestframework==3.14.0
mongoengine==0.27.0
PyJWT==2.8.0
bcrypt==4.1.1
python-dotenv==1.0.0
resend==0.6.0
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
daphne==4.0.0
Pillow>=10.0.0
requests==2.31.0
django-cors-headers==4.3.1
gunicorn==21.2.0

